Refactor to another file
